Renamed inference to pipeline and updated references
