+++ colorizer/pipeline.py
